Implement better filter framework, add lp filter
Filters work pretty much the same way as ADSR and oscs do. Separate
parameters for instruments, and local modifiable state per each channel.

Remove the old lp filter and add a new one using the framework. Note
that signals go from -1 to 1 and they must be divided by 2 to WAR the
overflow issue, same as in the ADSR.

The instrument's lp cutoff frequency can be tuned from the panel's pot.

Change the trivial saw to dpw saw in the instrument.