Velocity support for synth channels
Add a velocity field for the channels so that playing MIDI files with
a wide range of note velocities sounds correct. Make the keypad buttons
play the notes very silently for demoing the velocity effect.

For some reason, sending three words of data with an interrupt control
afterwards gets the whole system stuck! Workaround with bit magic.

Also increase the channel capacity to 63 for safety.