Sketch the filter coef tayloring with gnuplot
LFOs need to change the oscillators' frequencies and filters'
coefficients. Oscillator freq coefficients are in LUTs and can be
interpolated, but filter cutoff freq based coefficients are usually
needed only at compile time unless changed.

Fool around with taylor series because the coefficient formula is nasty
(= contains a division). Second order taylor error seems to be small.

Load in gnuplot with the 'load' command.

TODO: could a LUT for filt coefs help? It would need to be pretty big if
we'd compute accurate values for every possible cutoff frequency.