ADSR lfo for filter cutoff, clearer LFO structure
Add one instrument with ADSR controlled filter cutoff frequency. The LFO
effect means just modifying the filter cutoff with

 c(f) = c(a) + k * (f - a)

where k can be precalculated (lfo amplitude), and (f - a) is just the
lfo output; it's used to simply nudge the frequency slightly away from
the original position. This 1st order taylor is slightly inaccurate but
seems to sound good enough. This is the same for each LFO'd instrument,
so a macro.